Improve script to setup users and roles
