+++ x-pack/plugins/apm/ftr_e2e/cypress_start.ts
@@ -9,7 +9,7 @@ import Url from 'url';
 import cypress from 'cypress';
 import { FtrProviderContext } from './ftr_provider_context';
 import archives_metadata from './cypress/fixtures/es_archiver/archives_metadata';
-import { createKibanaUserRole } from '../scripts/kibana-security/create_kibana_user_role';
+import { createApmUsersAndRoles } from '../scripts/create-apm-users-and-roles/create_apm_users_and_roles';
 
 export function cypressRunTests(spec?: string) {
   return async ({ getService }: FtrProviderContext) => {
@@ -47,7 +47,7 @@ async function cypressStart(
   });
 
   // Creates APM users
-  await createKibanaUserRole({
+  await createApmUsersAndRoles({
     elasticsearch: {
       username: config.get('servers.elasticsearch.username'),
       password: config.get('servers.elasticsearch.password'),
