+++ x-pack/plugins/apm/scripts/create-apm-users-and-roles/helpers/call_kibana.ts
@@ -6,46 +6,51 @@
  */
 import axios, { AxiosRequestConfig, AxiosError } from 'axios';
 import { once } from 'lodash';
-import { Elasticsearch } from './create_kibana_user_role';
+import { Elasticsearch, Kibana } from '../create_apm_users_and_roles';
 
 export async function callKibana<T>({
   elasticsearch,
-  kibanaHostname,
+  kibana,
   options,
 }: {
   elasticsearch: Elasticsearch;
-  kibanaHostname: string;
+  kibana: Kibana;
   options: AxiosRequestConfig;
 }): Promise<T> {
-  const kibanaBasePath = await getKibanaBasePath({ kibanaHostname });
+  const baseUrl = await getBaseUrl(kibana.hostname);
   const { username, password } = elasticsearch;
 
   const { data } = await axios.request({
     ...options,
-    baseURL: kibanaHostname + kibanaBasePath,
+    baseURL: baseUrl,
     auth: { username, password },
     headers: { 'kbn-xsrf': 'true', ...options.headers },
   });
   return data;
 }
 
-const getKibanaBasePath = once(
-  async ({ kibanaHostname }: { kibanaHostname: string }) => {
-    try {
-      await axios.request({ url: kibanaHostname, maxRedirects: 0 });
-    } catch (e) {
-      if (isAxiosError(e)) {
-        const location = e.response?.headers?.location;
-        const isBasePath = RegExp(/^\/\w{3}$/).test(location);
-        return isBasePath ? location : '';
-      }
-
-      throw e;
+const getBaseUrl = once(async (kibanaHostname: string) => {
+  try {
+    await axios.request({ url: kibanaHostname, maxRedirects: 0 });
+  } catch (e) {
+    if (isAxiosError(e)) {
+      const location = e.response?.headers?.location;
+      const hasBasePath = RegExp(/^\/\w{3}$/).test(location);
+      const basePath = hasBasePath ? location : '';
+      return `${kibanaHostname}${basePath}`;
     }
-    return '';
+
+    throw e;
   }
-);
+  return kibanaHostname;
+});
 
 export function isAxiosError(e: AxiosError | Error): e is AxiosError {
   return 'isAxiosError' in e;
+}
+
+export class AbortError extends Error {
+  constructor(message: string) {
+    super(message);
+  }
 }
