+++ x-pack/plugins/apm/scripts/create-apm-users-and-roles/roles/power_user_role.ts
