+++ x-pack/plugins/apm/scripts/create-apm-users-and-roles/roles/read_only_user_role.ts
